docker: switch to traefik 2.8
Traefik 1.7 could not use multiple endpoints with the same certificate
(or at least I couldn't figure out how to make it work) but 2.8 works
easily with it.  Traefik 2.8 is actively maintained and it also
enables developer mode for us.

This also eliminates the need for config/traefik entirely.  There is
no configuration required beyond setting ACME_EMAIL and ACME_SERVER.